requirements.txt
amqp==2.5.0
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
Django==2.1.5
django-appconf==1.0.3
django-celery-email==2.0.2
djangorestframework==3.9.4
isort==4.3.4
kombu==4.6.3
pytz==2018.9
redis==3.2.0
six==1.12.0
vine==1.3.0
